+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # vsd Changelog (DD/MM/YYYY)
 
-## 0.1.1
+## 0.1.2 (09/07/2022)
 
 Features:
 
 - New `--build` flag.
-- `srt` subtitles collection with `--collect` flag.
+- `.srt` subtitles collection with `--collect` flag.
+
+Bug Fixes:
+
+- Not intercepting requests before navigating to website when using `--capture` and `--collect` flags.
 
 ## 0.1.0 (22/06/2022)